Moves shared HMS smoke type to top level module types
