+++ lib/hms_smoke/fetch_smoke/mod.ts
@@ -1,4 +1,3 @@
 export { fetchHMSSmoke } from "./fetchers.ts";
 export { getHMSSmokeUrl } from "./request_builders.ts";
 export { getHMSSmoke } from "./requests.ts";
-export * from "./types.ts";
